refactor: move common types into common.ts
